Fix missing environment variables for Wi-Fi
Add Wi-Fi override variables to mqtt_cfg.h. Since they are used
in the delegate Wi-Fi interface compilation might fail if they
are not defined even if Wi-Fi is not managed by the MQTT library.

Move WiFi.h include from header to cpp file, as it is not used
in the interface.